Rename b2 module to benchmark
