+++ src/gretel_trainer/benchmark/executor.py
@@ -6,8 +6,8 @@ from typing import Optional, Protocol
 from gretel_client.projects.models import Model
 from gretel_client.projects.projects import Project
 
-from gretel_trainer.b2.core import BenchmarkConfig, Dataset, log, run_out_path
-from gretel_trainer.b2.sdk_extras import create_evaluate_model, run_evaluate
+from gretel_trainer.benchmark.core import BenchmarkConfig, Dataset, log, run_out_path
+from gretel_trainer.benchmark.sdk_extras import create_evaluate_model, run_evaluate
 
 
 class Status(str, Enum):
